% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_package.r
\name{PTCMMLE}
\alias{PTCMMLE}
\title{different hazard-based regression structures (baseline, AFT, AH, PH, GH)}
\usage{
PTCMMLE(
  init,
  times,
  status,
  hstr = NULL,
  dist = NULL,
  des_theta = NULL,
  des_t = NULL,
  des_h = NULL,
  des = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
under the parameterisation (log(scale), log(shape1), log(shape2), alpha, eta, beta) for scale-shape1-shape2 models or
(mu, log(scale), alpha, eta, beta) for log-location scale models.}

\item{times}{: times to event}

\item{status}{: vital status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{hstr}{: hazard structure:
No covariates ("baseline"),
Accelerated Faiulre Time ("AFT"),
Proportional Hazards ("PH"),
Accelerated Hazards ("AH"),
General Hazards ("GH")
*GH is not available with Weibull "dist"}

\item{dist}{: distribution for the baseline hazard:
Power Generalised Weibull ("PGW")
Generalised Gamma ("GenGamma"))
Exponentiated Weibull ("EW")
Weibull ("Weibull")
Gamma ("Gamma")
LogNormal ("LogNormal")
LogLogistic ("LogLogistic")}

\item{des_theta}{: Design matrix for cure proportion parameter}

\item{des_t}{: Design matrix for time-level effects (GH)}

\item{des_h}{: Design matrix for hazard level effects (GH)}

\item{des}{: Design matrix for AFT, AH, and PH models}

\item{method}{: "nlminb" or optimisation method to be used in optim (see ?optim)}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
It returns the output from optim or nlminb for the selected model and the negative log likelihood function
}
\description{
different hazard-based regression structures (baseline, AFT, AH, PH, GH)
}
