% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_package.r
\name{simPTCMGH}
\alias{simPTCMGH}
\title{simPTCMGH function: simulates from a promotion-time cure model with 
different hazard-based regression structures (baseline, AFT, AH, PH, GH)}
\usage{
simPTCMGH(
  n,
  seed = 123,
  hstr = NULL,
  dist = NULL,
  des_theta = NULL,
  des_t = NULL,
  des_h = NULL,
  des = NULL,
  par_base = NULL,
  alpha0 = NULL,
  alpha = NULL,
  beta_t = NULL,
  beta_h = NULL,
  beta = NULL
)
}
\arguments{
\item{n}{: sample size (number of individuals)}

\item{seed}{: seed for simulation}

\item{hstr}{: hazard structure (baseline, AFT, AH, PH, GH)}

\item{dist}{: baseline hazard distribution}

\item{des_theta}{: Design matrix for cure proportion parameter}

\item{des_t}{: Design matrix for time-level effects (GH)}

\item{des_h}{: Design matrix for hazard level effects (GH)}

\item{des}{: Design matrix for AFT, AH, and PH models}

\item{par_base}{:  parameters of the baseline hazard}

\item{alpha}{:  parameters of the regression model on the cure proportion parameter}

\item{beta_t}{: regression parameters multiplying the time scale for GH model}

\item{beta_h}{: regression parameters multiplying the hazard for GH model}

\item{beta}{: regression parameters for AFT, PH, and AH models}
}
\value{
a vector containing the simulated times to event
}
\description{
simPTCMGH function: simulates from a promotion-time cure model with 
different hazard-based regression structures (baseline, AFT, AH, PH, GH)
}
